[fix] immutable: don't recreate if nothing changed; Provide data into creation strategy for that reason

+++ src/Creation/ImmutableStrategy.php
@@ -11,13 +11,16 @@ class ImmutableStrategy implements CreationStrategy
 {
   use GetManager;
 
-  private $idMap = [];
+  /** @var array */
+  protected $serviceFields;
+  private   $idMap = [];
 
-  public function __construct(ManagerRegistry $registry)
+  public function __construct(ManagerRegistry $registry, $serviceFields = [])
   {
-    $this->registry = $registry;
+    $this->registry      = $registry;
+    $this->serviceFields = $serviceFields;
   }
-  public function create(ClassMetadata $meta, ?array $id, $entity = null)
+  public function create(ClassMetadata $meta, ?array $id, $entity = null, $data = null)
   {
     $className = $meta->getName();
     $em        = $this->getManager($className);
@@ -25,10 +28,14 @@ class ImmutableStrategy implements CreationStrategy
       $key    = implode('|', $id);
       $entity = $entity ?? $this->idMap[$className][$key] ?? $em->getRepository($className)->findOneBy($id);
     }
-    if (isset($entity)) {
+    if (isset($entity) && $this->hasEntityChanged($entity, $id, $data)) {
       $em->remove($entity);
+      $entity = null;
     }
-    $entity = $meta->newInstance();
+    if (!$entity) {
+      $entity = $meta->newInstance();
+    }
+
     if (isset($key)) {
       $this->idMap[$className][$key] = $entity;
     }
@@ -38,5 +45,26 @@ class ImmutableStrategy implements CreationStrategy
   public function clearState(): void
   {
     $this->idMap = [];
+  }
+  private function hasEntityChanged($entity, array $id, array $data): bool
+  {
+    $fields = is_callable($this->serviceFields) ? ($this->serviceFields)($entity) : $this->serviceFields;
+
+    $entityData      = get_object_vars($entity);
+    $serviceFieldMap = array_flip($fields);
+    $changed         = false;
+    foreach ($entityData as $field => $value) {
+      if (array_key_exists($field, $id) || array_key_exists($field, $serviceFieldMap)) {
+        continue;
+      }
+      if (strpos($field, '_') === 0) {
+        continue;
+      }
+      if (($data[$field] ?? null) !== ($entity->$field ?? null)) {
+        $changed = true;
+        break;
+      }
+    }
+    return $changed;
   }
 }
